# script/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "ABAC"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/resources": {
      "get": {
        "summary": "List all resources",
        "operationId": "list",
        "tags": [
          "resources"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "page number of the list request",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "limit of number of results fetched from the list request",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of resources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pagingMetadata": {
                      "$ref": "#/components/schemas/PagingMetadata"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Resources"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a resource",
        "operationId": "create",
        "tags": [
          "resources"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Resource"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "resource created successfully"
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/resources/{resourceId}": {
      "get": {
        "summary": "fetch resource by its id",
        "operationId": "getById",
        "tags": [
          "resources"
        ],
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "description": "The id of the resource to fetch",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull response for the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "To update a resource by id",
        "operationId": "updateById",
        "tags": [
          "resources"
        ],
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "description": "The id of the resource to fetch",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "$ref": "#/components/schemas/Resource"
                  },
                  "fieldMask": {
                    "$ref": "#/components/schemas/FieldMask"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "resource updated successfully"
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "To delete a resource by id",
        "operationId": "deleteById",
        "tags": [
          "resources"
        ],
        "parameters": [
          {
            "name": "resourceId",
            "in": "path",
            "required": true,
            "description": "The id of the resource to fetch",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "resource deleted successfully"
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Resource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "policyId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "deleted": {
            "type": "string",
            "format": "date-time"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Resources": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Resource"
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "example": 404,
            "format": "int32"
          },
          "message": {
            "type": "string",
            "example": "Not found"
          }
        }
      },
      "FieldMask": {
        "type": "object",
        "properties": {
          "paths": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ownerId"
            }
          }
        }
      },
      "PagingMetadata": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 100,
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "example": 10,
            "format": "int64"
          },
          "pageNumber": {
            "type": "integer",
            "example": 1,
            "format": "int64"
          }
        }
      }
    }
  }
}